requirements.txt
shiny>=0.6.0
shinyswatch>=0.4.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
plotly>=5.0.0
python-multipart>=0.0.6
cryptography>=41.0.0
